posts-service/requirements.txt
Django~=6.0
requests~=2.32.5
djangorestframework~=3.16.1
django-filter~=25.2
kafka~=1.3.5
psycopg2==2.9.11
djangorestframework-simplejwt==5.5.1
pytest~=9.0.1
model-bakery~=1.20.5